Expand concept of fields with defaults, and re-think timestamp
The timestamp is now encoded as an offset from the previous
record, and may default to 0 allowing it to be encoded as
simply a TAB if the message happens at the same moment as
the previous.  Also name the column timestamp_step_hex to
help self-document it, and leave room for future extensions.
Also make the public accessor always return the unix epoch
and hide the implementation details of offsets or steps.