Document and simplify the tsv0 format
I had new ideas for the tsv0 format which make it even less
simple, so I decided there should be a tsv0 format which is
the simplest of all, and then the default will be tsv1 with
the optimizations.  tsv0 will advertise not needing a library
and tsv1 will advertise being easily readable and smaller,
but needing a simple library to help decode it.